util: Make pfconfig generic
